% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probation-caseload.R
\name{probationCaseload}
\alias{probationCaseload}
\title{Convert a Probation Caseload CSV to Tidy format}
\usage{
probationCaseload(file)
}
\arguments{
\item{file}{Path to the input file (CSV)}
}
\value{
A data frame of the input file in Tidy format
}
\description{
Convert a Probation Caseload CSV to Tidy format
}
