% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probation-starts.R
\name{probationStarts}
\alias{probationStarts}
\title{Convert a Probation Starts CSV to Tidy format data frame}
\usage{
probationStarts(file)
}
\arguments{
\item{file}{Path to the input file (CSV)}
}
\value{
A data frame of the input file in Tidy format
}
\description{
Convert a Probation Starts CSV to Tidy format data frame
}
